Cursor now working as expected.
Added logic to create a sprite when taking a picture. Need to make sure we're making the image at the position of the cursor. Since the logic is being called from the cursor, I expect this to be the case.It might just be a blank image, though. Testing needed.

Trying to add logic to the WaterMap tilemap, so we can detect when things are walking in the water. We might need to programatically change out water tiles for Area2D scenes, assuming that doesn't hit performance too much. If that works, we'll also try that with the grass and bushes.